fix(webhook): validate match rule in nat firewallconfiguration
